hw6: finished collect_relationships

+++ hw6/submission_template/src/collect_relationships.py
@@ -5,6 +5,7 @@ import json
 import requests
 import shutil
 
+from json import JSONDecodeError
 from pathlib import Path
 
 
@@ -13,6 +14,13 @@ ROOT_DIR = Path(__file__).parents[1].absolute()
 
 
 def populate_cache_dir(config_dict):
+    
+    print(f"Checking cache at {config_dict['cache_dir']}...")
+    if not os.path.exists(config_dict['cache_dir']):
+        print(f"No cache found, creating directory...")
+    else:
+        print(f"Cache exists, will double check to make sure there is a cached file for each target person...")
+
     cache_loc = os.path.join(ROOT_DIR, config_dict['cache_dir'])
 
     p = Path(os.path.join(cache_loc))
@@ -20,11 +28,14 @@ def populate_cache_dir(config_dict):
     
     for person in config_dict['target_people']: 
         person_page_fname = os.path.join(cache_loc, person)
-        print(f"Downloading whodatedwho info for {person} to location {person_page_fname}...")
-        with requests.get(BASE_URL + person, stream=True) as r:
-            with open(person_page_fname, 'wb') as f:
-                shutil.copyfileobj(r.raw, f)
-        print(f"Download successful for {person} to {person_page_fname}!")
+        if not os.path.exists(person_page_fname):
+            print(f"Downloading whodatedwho info for {person} to location {person_page_fname}...")
+            with requests.get(BASE_URL + person, stream=True) as r:
+                with open(person_page_fname, 'wb') as f:
+                    shutil.copyfileobj(r.raw, f)
+            print(f"Download successful for {person} to {person_page_fname}!")
+        else:
+            print(f"Cache already exists for {person} in cache dir {cache_loc}!")
 
     return None
 
@@ -35,19 +46,22 @@ def extract_page_files_from_cache(config_dict):
 
 
 def get_relationships(person_page_fname):
-
-
-
     soup = bs4.BeautifulSoup(open(person_page_fname, 'r'), 'html.parser')
     data = soup.select("[type='application/ld+json']")[0]
-    oJson = json.loads(data.text)["itemListElement"]
-    numRelations = len(oJson)
-    results = []
 
-    for product in oJson:
-        results.append(product['item']['name'])
+    result = [] 
+    try:
+        relationships_list = json.loads(data.text)
+    except JSONDecodeError as e:
+        print(f"Error: malformed relationship json for {person_page_fname}, will report empty list for target")
+        return result
 
-    return results
+    try:
+        result = [relationship['item']['name'] for relationship in relationships_list["itemListElement"]]
+    except KeyError as e:
+        print(f"Warning: no relationship information for {person_page_fname}, will report empty list for target") 
+
+    return result
 
 
 def main(args):
@@ -56,19 +70,13 @@ def main(args):
     with open(args.configFile, 'r') as fh:
         config_dict = json.load(fh)
 
-    # if cache_dir doesn't exist, create it (dl file for each target person)
-    print(f"Checking cache at {config_dict['cache_dir']}...")
-    if not os.path.exists(config_dict['cache_dir']):
-        print(f"No cache found at {config_dict['cache_dir']}, populating cache by downloading from {BASE_URL}...")
-        populate_cache_dir(config_dict)
-    print(f"Cache found, extracting relationship data...")
+    populate_cache_dir(config_dict)
+    print(f"Cache ready, extracting relationship data...")
 
     person_file_map = extract_page_files_from_cache(config_dict)
+    output = {person: get_relationships(person_page_fname) for person, person_page_fname in person_file_map.items()}
 
-    # for each target person, read info from cache.
-    output = { person: get_relationships(person_page_fname) for person, person_page_fname in person_file_map.items() }
-
-    # publish dict to json
+    print(f"Writing relationship data to {args.outputFile}...")
     with open(args.outputFile, 'w') as fh:
         json.dump(output, fh)
